refactor and mock pdf preview
